refactored backend for organization
